+++ cicada/tree_grammar_mutable.cpp
@@ -298,7 +298,7 @@ namespace cicada
 	    features[feature_names[feature]] = fiter->second;
 	  else {
 	    // default name!
-	    if (feature >= feature_names_default.size())
+	    if (feature >= static_cast<int>(feature_names_default.size()))
 	      feature_names_default.resize(feature + 1);
 	    if (feature_names_default[feature].empty())
 	      feature_names_default[feature] = "rule-table-" + boost::lexical_cast<std::string>(feature);
@@ -321,7 +321,7 @@ namespace cicada
 	    attributes[attribute_names[attribute]] = aiter->second;
 	  else {
 	    // default name!
-	    if (attribute >= attribute_names_default.size())
+	    if (attribute >= static_cast<int>(attribute_names_default.size()))
 	      attribute_names_default.resize(attribute + 1);
 	    if (attribute_names_default[attribute].empty())
 	      attribute_names_default[attribute] = "rule-table-" + boost::lexical_cast<std::string>(attribute);
@@ -389,7 +389,7 @@ namespace cicada
     for (scores_parsed_type::const_iterator fiter = boost::fusion::get<0>(scores_attrs).begin(); fiter != fiter_end; ++ fiter)
       if (fiter->first.empty()) {
 	// default name!
-	if (feature >= feature_names_default.size())
+	if (feature >= static_cast<int>(feature_names_default.size()))
 	  feature_names_default.resize(feature + 1);
 	if (feature_names_default[feature].empty())
 	  feature_names_default[feature] = "rule-table-" + boost::lexical_cast<std::string>(feature);
@@ -405,7 +405,7 @@ namespace cicada
     for (attrs_parsed_type::const_iterator aiter = boost::fusion::get<1>(scores_attrs).begin(); aiter != aiter_end; ++ aiter)
       if (aiter->first.empty()) {
 	// default name!
-	if (attribute >= attribute_names_default.size())
+	if (attribute >= static_cast<int>(attribute_names_default.size()))
 	  attribute_names_default.resize(attribute + 1);
 	if (attribute_names_default[attribute].empty())
 	  attribute_names_default[attribute] = "rule-table-" + boost::lexical_cast<std::string>(attribute);
